Get taxonomy ranks using NCBI Taxonomy Tax ID

Recover complete taxonomy for a given NCBI Taxonomy Tax ID

Usage

get_tax_by_taxID(
  organism_taxID,
  parse_result = TRUE,
  total_cores = 1,
  show_command = TRUE
)

Arguments

parse_result

Should the taxonomy be returned as the efetch returns it or should it be parsed into a tibble.

total_cores

Number of threads to use. Defaults to 1.

show_command

Should the efetch command on command line be printed for debugging. Defaults to TRUE.

taxIDs

Vector of NCBI Taxonomy Tax ID to retrieve taxonomy for.

Value

A tibble with all the taxonomic ranks for the corresponding taxID.

Examples

blast_res <- BLASTr::run_blast(
  asv = "CTAGCCATAAACTTAAATGAAGCTATACTAAACTCGTTCGCCAG
    AGTACTACAAGCGAAAGCTTAAAACTCATAGGACTTGGCGGTGTTTCAGACCCAC",
  db_path = "/data/databases/nt/nt",
  perc_id = 80,
  num_thread = 1,
  perc_qcov_hsp = 80,
  num_alignments = 2
)
#> Error in BLASTr:::check_bin(blast_type): ✖ blastn is not available on PATH.
#> ℹ Please install it and try again.